gamemode: add new portal for GameMode access
Add a new portal so that GameMode service can be accessed from
within the sandbox. GameMode expects to get the process id (pid)
of the game so it can:
  - track the lifetime of the game
  - modify process parameters (nice level, scheduler policy, ...)
  - resolve the executable name and trigger configured actions

Since processes inside the sandbox run in a separate pid namespace
sending the client pid obtained via getpid(2) won't work. Getting
the pid via the D-Bus connection will result in getting the pid of
the dbus proxy. According to pid_namespaces(7) process ids will be
translated by the kernel when they are passed over a unix domain
socket. But passing a connected unix socket via dbus messages and
then obtaining the credentials for that socket (pair) will also
resolve the pid to the dbus proxy. Therefore we nest one level more
and send a socket of a socket pair via a connected socket that was
passed via dbus. This additional level of indirection then results
in the correct credentials.